+++ common/src/main/scala/explore/components/ui/ExploreStyles.scala
@@ -13,20 +13,20 @@ object ExploreStyles:
 
   val HideReusability = Css("hide-reusability")
 
-  val Tile: Css                    = Css("explore-tile")
-  val TileTitle: Css               = Css("explore-tile-title")
-  val TileTitleText: Css           = Css("explore-tile-title-text")
-  val TileTitleMenu: Css           = Css("explore-tile-title-menu")
+  val Tile: Css                        = Css("explore-tile")
+  val TileTitle: Css                   = Css("explore-tile-title")
+  val TileTitleText: Css               = Css("explore-tile-title-text")
+  val TileTitleMenu: Css               = Css("explore-tile-title-menu")
   // val TileTitleStrip: Css          = Css("explore-tile-title-strip")
-  val TileDraggable: Css           = Css("explore-tile-draggable")
-  val TileTitleControlArea: Css    = Css("explore-tile-title-control-area")
-  val JustifiedEndTileControl: Css = Css("justified-end-tile-control")
-  val TileControlButtons: Css      = Css("explore-tile-control-buttons")
-  val TileBackButton: Css          = Css("tile-back-button")
-  val TileBody: Css                = Css("explore-tile-body")
-  val TileButton: Css              = Css("explore-tile-button")
-  val TileControl: Css             = Css("explore-tile-control")
-  val TileStateButton: Css         = Css("explore-tile-state-button")
+  val TileDraggable: Css               = Css("explore-tile-draggable")
+  val TileTitleControlArea: Css        = Css("explore-tile-title-control-area")
+  val TileTitleConstraintSelector: Css = Css("explore-tile-title-constraint-selector")
+  val TileControlButtons: Css          = Css("explore-tile-control-buttons")
+  val TileBackButton: Css              = Css("tile-back-button")
+  val TileBody: Css                    = Css("explore-tile-body")
+  val TileButton: Css                  = Css("explore-tile-button")
+  val TileControl: Css                 = Css("explore-tile-control")
+  val TileStateButton: Css             = Css("explore-tile-state-button")
 
   val Accented: Css    = Css("explore-accented")
   val TextPlain: Css   = Css("explore-text-plain")
